refactor: ReviewCard 폴더 위치 이동 및 디자인 수정
