+++ frontend/src/components/common/ReviewCard/StrengthKeywordList/index.tsx
