+++ frontend/src/components/common/ReviewCard/StrengthKeywordList/styles.ts
